+++ benches/effects.rs
@@ -1,7 +1,6 @@
 use std::time::Duration;
 
 use divan::Bencher;
-use rodio::source::AutomaticGainControlSettings;
 use rodio::Source;
 
 mod shared;
@@ -48,7 +47,7 @@ fn amplify(bencher: Bencher) {
 fn agc_enabled(bencher: Bencher) {
     bencher.with_inputs(music_wav).bench_values(|source| {
         source
-            .automatic_gain_control(AutomaticGainControlSettings::default())
+            .automatic_gain_control(Default::default())
             .for_each(divan::black_box_drop)
     })
 }
@@ -58,8 +57,7 @@ fn agc_enabled(bencher: Bencher) {
 fn agc_disabled(bencher: Bencher) {
     bencher.with_inputs(music_wav).bench_values(|source| {
         // Create the AGC source
-        let amplified_source =
-            source.automatic_gain_control(AutomaticGainControlSettings::default());
+        let amplified_source = source.automatic_gain_control(Default::default());
 
         // Get the control handle and disable AGC
         let agc_control = amplified_source.get_agc_control();