+++ src/source/skip.rs
@@ -2,9 +2,8 @@ use std::time::Duration;
 
 use super::SeekError;
 use crate::common::{ChannelCount, SampleRate};
+use crate::math::NANOS_PER_SEC;
 use crate::Source;
-
-const NS_PER_SECOND: u128 = 1_000_000_000;
 
 /// Internal function that builds a `SkipDuration` object.
 pub fn skip_duration<I>(mut input: I, duration: Duration) -> SkipDuration<I>
@@ -42,7 +41,7 @@ where
         let sample_rate = input.sample_rate().get() as u128;
         let channels = input.channels().get() as u128;
 
-        let samples_per_channel = duration.as_nanos() * sample_rate / NS_PER_SECOND;
+        let samples_per_channel = duration.as_nanos() * sample_rate / NANOS_PER_SEC as u128;
         let samples_to_skip: u128 = samples_per_channel * channels;
 
         // Check if we need to skip only part of the current span.
@@ -52,7 +51,7 @@ where
         }
 
         duration -= Duration::from_nanos(
-            (NS_PER_SECOND * span_len as u128 / channels / sample_rate) as u64,
+            (NANOS_PER_SEC as u128 * span_len as u128 / channels / sample_rate) as u64,
         );
         skip_samples(input, span_len);
     }
@@ -65,7 +64,7 @@ where
     I: Source,
 {
     let samples_per_channel: u128 =
-        duration.as_nanos() * input.sample_rate().get() as u128 / NS_PER_SECOND;
+        duration.as_nanos() * input.sample_rate().get() as u128 / NANOS_PER_SEC as u128;
     let samples_to_skip: u128 = samples_per_channel * input.channels().get() as u128;
 
     skip_samples(input, samples_to_skip as usize);
@@ -171,6 +170,8 @@ mod tests {
     use crate::common::{ChannelCount, SampleRate};
     use crate::math::nz;
     use crate::source::Source;
+    use crate::Sample;
+    use dasp_sample::Sample as DaspSample;
 
     fn test_skip_duration_samples_left(
         channels: ChannelCount,
@@ -180,7 +181,7 @@ mod tests {
     ) {
         let buf_len = (sample_rate.get() * channels.get() as u32 * seconds) as usize;
         assert!(buf_len < 10 * 1024 * 1024);
-        let data: Vec<f32> = vec![0f32; buf_len];
+        let data: Vec<Sample> = vec![Sample::EQUILIBRIUM; buf_len];
         let test_buffer = SamplesBuffer::new(channels, sample_rate, data);
         let seconds_left = seconds.saturating_sub(seconds_to_skip);